+++ internal/exec/exec.go
@@ -86,6 +86,13 @@ func RunnerWithCachePath(cachePath string) RunnerOption {
 	}
 }
 
+// RunnerWithConfigData returns a RunnerOption that uses the given config path.
+func RunnerWithConfigData(configData string) RunnerOption {
+	return func(runner *runner) {
+		runner.configData = configData
+	}
+}
+
 // RunnerWithJSON returns a RunnerOption that will print failures as JSON.
 func RunnerWithJSON() RunnerOption {
 	return func(runner *runner) {
@@ -98,6 +105,20 @@ func RunnerWithJSON() RunnerOption {
 func RunnerWithPrintFields(printFields string) RunnerOption {
 	return func(runner *runner) {
 		runner.printFields = printFields
+	}
+}
+
+// RunnerWithProtocBinPath returns a RunnerOption that uses the given protoc binary path.
+func RunnerWithProtocBinPath(protocBinPath string) RunnerOption {
+	return func(runner *runner) {
+		runner.protocBinPath = protocBinPath
+	}
+}
+
+// RunnerWithProtocWKTPath returns a RunnerOption that uses the given path to include the well-known types.
+func RunnerWithProtocWKTPath(protocWKTPath string) RunnerOption {
+	return func(runner *runner) {
+		runner.protocWKTPath = protocWKTPath
 	}
 }